Don't add OpenSSL as a dependency on Windows
This is awkward but OpenSSL is only used on Unix, and we only want to include it here so we can set the `vendored` feature in that case. We don't want it unconditionally because then it tries to build it on Windows which a) we don't want, and b) fails because we don't have Perl, because it's not the feaking 90s.